pyproject.toml
[tool.poetry]
name = "subscription-manager"
version = "2.0"
description = "Web application for managing newspaper subscriptions"
authors = ["Robin Bisping"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0"
django-compressor = "^2.4"
django-debug-toolbar = "^2.1"
django-import-export = "^2.0.1"
django-libsass = "^0.8"
django_cron = "^0.5.1"
django-dotenv = "^1.4.2"
psycopg2 = "^2.8.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
